// @types/StatusInvestNormalizedDataType.d.ts
export type StatusInvestNormalizedDataType = {
  TICKER: string;
  PRECO: number;
  DY: number;
  "P/L": number;
  "P/VP": number;
  "P/ATIVOS": number;
  "MARGEM BRUTA": number;
  "MARGEM EBIT": number;
  "MARG. LIQUIDA": number;
  "P/EBIT": number;
  "EV/EBIT": number;
  "DIVIDA LIQUIDA / EBIT": number;
  "DIV. LIQ. / PATRI.": number;
  PSR: number;
  "P/CAP. GIRO": number;
  "P. AT CIR. LIQ.": number;
  "LIQ. CORRENTE": number;
  ROE: number;
  ROA: number;
  ROIC: number;
  "PATRIMONIO / ATIVOS": number;
  "PASSIVOS / ATIVOS": number;
  "GIRO ATIVOS": number;
  "CAGR RECEITAS 5 ANOS": number;
  "CAGR LUCROS 5 ANOS": number;
  " LIQUIDEZ MEDIA DIARIA": number;
  " VPA": number;
  " LPA": number;
  " PEG Ratio": number;
  " VALOR DE MERCADO": number;
};